requirements.txt
boto3
connexion
gevent>=1.1b1
flowlogs_reader
stups-tokens
clickclick
